--- chat-plugins/badgeicons.js
@@ -1,38 +1,0 @@
-'use strict';
-/************************************************
- * Badge Icons *
- ************************************************/
-
-module.exports = {
-	"rpsmaster":"http://i.imgur.com/xpRhWgI.png",
-	"staff":"http://i.imgur.com/EqwEcfD.gif",
-	"Nolife Master":"http://i.imgur.com/pKxK9pv.pngx",
-	"Collector":"http://i.imgur.com/a26Qbmi.png",
-	"Ace":"http://i.imgur.com/uujwykI.png",
-	"Annoying":"http://i.imgur.com/GdnPYJi.png",
-	"Mono Pro":"http://i.imgur.com/R0VISSU.png",
-	"Champ":"http://i.imgur.com/ttGM1SR.png",
-	"LUCKY MOFO":"http://i.imgur.com/KH8m6Rg.png",
-	"High Roller":"http://i.imgur.com/1AulGHm.png",
-	"Achievement Unlocked!":"http://i.imgur.com/YRgMD4s.png",
-	"Persistent!":"http://i.imgur.com/C6gknys.png",
-	"Active Poster!":"http://i.imgur.com/GuGCyrY.png",
-	"Float Badge":"http://i.imgur.com/1vdYBhD.png",
-	"Commie Badge":"http://i.imgur.com/5jzkxd3.gif",
-	"Cannon Badge":"http://i.imgur.com/QPofWrT.png",
-	"Pixel Badge":"http://i.imgur.com/ymSJ1MQ.png",
-	"Neo Badge":"http://i.imgur.com/Zs9VGbg.png",
-	"Century Badge":"http://i.imgur.com/dPR2mBM.png",
-	"Neski Badge":"http://i.imgur.com/ZBRW9CZ.png",
-	"Arken Badge":"http://i.imgur.com/KvELNsF.png",
-	"Alpha Badge":"http://i.imgur.com/JfflYtk.png",
-	"Gryphon Badge":"http://i.imgur.com/crtS5z2.png",
-	"fender Badge":"http://i.imgur.com/s16lNTS.gif",
-	"Lucario Badge":"http://i.imgur.com/yfSYq1w.gif",
-	"weeb":"http://i.imgur.com/XJmjJDE.png",
-	"Meme Lord":"http://i.imgur.com/4nLKC4V.png",
-	"Cute Fox":"http://i.imgur.com/xgN76US.png",
-	"vip":"http://i.imgur.com/7heNNTP.png",
-	"Donor of the month":"http://i.imgur.com/EHx2ozm.gif",
-	"Neptune badge":"http://i.imgur.com/c0eYyOf.png",
-};
